made discussion out of methods.py to new discussion.py
